# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000"
        },
        {
            "url": "http://swagger.local"
        }
    ],
    "paths": {
...
        "/api/listActivities": {
            "get": {
                "tags": [
                    "Lista Actividades"
                ],
                "summary": "Mostrar lista de todos los tipos de actividades",
                "operationId": "App\\Http\\Controllers\\Api\\ListActivitiesController::get",
                "responses": {
                    "200": {
                        "description": "Mostrar todos los tipos de actividades."
                    },
                    "204": {
                        "description": "No hay resultados que mostrar."
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                }
            }
        },
        "/api/listActivities/{id}": {
            "get": {
                "tags": [
                    "Lista Actividades"
                ],
                "summary": "Mostrar datos de tipo de actividad por id",
                "operationId": "App\\Http\\Controllers\\Api\\ListActivitiesController::getByid",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id especifico a retornar",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Mostrar todos los tipos de actividades."
                    },
                    "204": {
                        "description": "No hay resultados que mostrar."
                    },
                    "400": {
                        "description": "Request mal mandado."
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                }
            }
        },
        "/api/enterprises": {
            "get": {
                "tags": [
                    "Listar empresas "
                ],
                "summary": "Muestra todas las empresas asociadas a sig",
                "operationId": "App\\Http\\Controllers\\Api\\ListEnterprisesController::getEnterprises",
                "responses": {
                    "200": {
                        "description": "Mostrar todos las empresas."
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                }
            }
        },
        "/api/customers": {
            "get": {
                "tags": [
                    "Listar empresas "
                ],
                "summary": "Muestra todos los clientes",
                "operationId": "App\\Http\\Controllers\\Api\\ListEnterprisesController::getCustomers",
                "responses": {
                    "200": {
                        "description": "Mostrar todos las clientes."
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                }
            }
        },
        "/api/users": {
            "get": {
                "tags": [
                    "usuario"
                ],
                "summary": "Mostrar usuarios",
                "operationId": "App\\Http\\Controllers\\Api\\UserController::get",
                "responses": {
                    "200": {
                        "description": "Usuarios encontrados."
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                }
            }
        },
        "/api/users/{id}": {
            "get": {
                "tags": [
                    "usuario"
                ],
                "summary": "Mostrar usuario por id",
                "operationId": "App\\Http\\Controllers\\Api\\UserController::getByid",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id del usuario",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Usuario encontrado."
                    },
                    "400": {
                        "description": "Request mal mandado."
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                }
            }
        }
    }
}